+++ src/main/java/janggi/domain/board/Position.java
@@ -1,4 +1,4 @@
-package janggi.board;
+package janggi.domain.board;
 
 import java.util.List;
 import java.util.Set;
