+++ src/main/java/janggi/view/OutputView.java
@@ -1,16 +1,16 @@
 package janggi.view;
 
-import janggi.Score;
-import janggi.Turn;
-import janggi.board.Position;
-import janggi.piece.Cannon;
-import janggi.piece.Elephant;
-import janggi.piece.Guard;
-import janggi.piece.Horse;
-import janggi.piece.King;
-import janggi.piece.Piece;
-import janggi.piece.Side;
-import janggi.piece.Soldier;
+import janggi.domain.board.Position;
+import janggi.domain.game.Score;
+import janggi.domain.game.Turn;
+import janggi.domain.piece.Cannon;
+import janggi.domain.piece.Elephant;
+import janggi.domain.piece.Guard;
+import janggi.domain.piece.Horse;
+import janggi.domain.piece.King;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.Side;
+import janggi.domain.piece.Soldier;
 import java.util.Map;
 
 public class OutputView {
