+++ src/test/java/janggi/board/BoardTest.java
@@ -5,13 +5,15 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import janggi.Score;
-import janggi.Turn;
-import janggi.piece.Cannon;
-import janggi.piece.Piece;
-import janggi.piece.Side;
-import janggi.piece.Soldier;
-import janggi.piece.Tank;
+import janggi.domain.board.Board;
+import janggi.domain.board.Position;
+import janggi.domain.game.Score;
+import janggi.domain.game.Turn;
+import janggi.domain.piece.Cannon;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.Side;
+import janggi.domain.piece.Soldier;
+import janggi.domain.piece.Tank;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
