+++ src/test/java/janggi/board/PositionTest.java
@@ -2,6 +2,7 @@ package janggi.board;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import janggi.domain.board.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
