+++ src/test/java/janggi/piece/ElephantTest.java
@@ -2,7 +2,11 @@ package janggi.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import janggi.board.Position;
+import janggi.domain.board.Position;
+import janggi.domain.piece.Elephant;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.Side;
+import janggi.domain.piece.Tank;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
