+++ src/test/java/janggi/piece/GuardTest.java
@@ -2,7 +2,9 @@ package janggi.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import janggi.board.Position;
+import janggi.domain.board.Position;
+import janggi.domain.piece.Guard;
+import janggi.domain.piece.Side;
 import java.util.Map;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
